backend/services/manipulate/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
-e file:///Users/jongwankim/Desktop/workspace/rby2/backend/packages/flat_buffers
    # via manipulate (pyproject.toml)
-e file:///Users/jongwankim/Desktop/workspace/rby2/backend/packages/rb_zenoh
    # via manipulate (pyproject.toml)
-e file:///Users/jongwankim/Desktop/workspace/rby2/backend/packages/utils
    # via manipulate (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
click==8.2.1
    # via uvicorn
eclipse-zenoh==1.5.0
    # via rb-zenoh
fastapi==0.116.1
    # via manipulate (pyproject.toml)
flatbuffers==25.2.10
    # via
    #   flat-buffers
    #   rb-zenoh
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
psutil==7.0.0
    # via
    #   rb-zenoh
    #   utils
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via
    #   manipulate (pyproject.toml)
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via manipulate (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
